+++ sqllin-dsl/src/commonMain/kotlin/com/ctrip/sqllin/dsl/DBEntity.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) 2022 Ctrip.com.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.ctrip.sqllin.dsl
+
+import kotlinx.serialization.KSerializer
+
+/**
+ * Base where clause property in SQL
+ * @author yaqiao
+ */
+
+@Deprecated(message = "The DBEntity hae benn deprecated, you don't need to implement this interface anymore")
+public fun interface DBEntity<T> {
+
+    public fun kSerializer(): KSerializer<T>
+}